requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
motor==3.3.2
pytesseract==0.3.10
Pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
scikit-learn==1.4.0
pdf2image
opencv-python
numpy
python-docx
openai


